feat: Created programmatic configuration for CLI options (#1887)
