+++ samples/unity-of-bugs/Assets/Resources/Sentry/SentryCliConfiguration.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8c8107e4a8950413eb741aa6d47c3b73
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
