+++ samples/unity-of-bugs/Assets/Scripts/SentryCliConfiguration.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: be9687ec81ccb4180b0ed8c055df62d8
